fix: list create funcionando, mas volta para a home sem a lista selecionada

+++ lib/service/queries/list_service.dart
@@ -43,7 +43,7 @@ class ListService {
     required String token,
     required String name,
     String? emoji,
-    int? maxSpend,
+    double? maxSpend,
   }) async {
     const String path = '/api/list';
